Revert "feat(ckg-api): npm test fix attempt"
This reverts commit 124126734073d4422471a0bf7488cdd42ae3d508.
