+++ express/scripts/templates.js
@@ -18,89 +18,9 @@ import {
   getLocale,
 } from './scripts.js';
 
-const endpoints = {
-  dev: {
-    cdn: '',
-    url: 'https://uss-templates-dev.adobe.io/uss/v3/query',
-    token: 'cd1823ed-0104-492f-ba91-25f4195d5f6c',
-  },
-  stage: {
-    cdn: 'https://www.stage.adobe.com/ax-uss-api/',
-    url: 'https://uss-templates-stage.adobe.io/uss/v3/query',
-    token: 'db7a3d14-5aaa-4a3d-99c3-52a0f0dbb459',
-    key: 'express-ckg-stage',
-  },
-  prod: {
-    cdn: '',
-    url: 'https://uss-templates-dev.adobe.io/uss/v3/query',
-    token: '2e0199f4-c4e2-4025-8229-df4ca5397605',
-  },
-};
-
-export default async function getData(env = '', data = {}) {
-  const endpoint = endpoints[env];
-  const response = await fetch(endpoint.url, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/vnd.adobe.search-request+json',
-      'x-api-key': endpoint.key,
-      'x-app-token': endpoint.token,
-    },
-    body: JSON.stringify(data),
-  });
-
-  if (response.ok) {
-    return response.json();
-  } else {
-    return response;
-  }
-}
-
-export async function fetchLInkListFromCKGApi(pageData) {
-  const params = new Proxy(new URLSearchParams(window.location.search), {
-    get: (searchParams, prop) => searchParams.get(prop),
-  });
-
-  if (pageData.ckgID || params.ckgid) {
-    const dataRaw = {
-      experienceId: 'templates-browse-v1',
-      locale: 'en_US',
-      queries: [
-        {
-          id: 'ccx-search-1',
-          start: 0,
-          limit: 40,
-          scope: {
-            entities: [
-              'HzTemplate',
-            ],
-          },
-          filters: [
-            {
-              categories: [
-                pageData.ckgID ?? params.ckgid,
-              ],
-            },
-          ],
-          facets: [
-            {
-              facet: 'categories',
-              limit: 10,
-            },
-          ],
-        },
-      ],
-    };
-
-    const env = getHelixEnv();
-    const result = await getData(env.name, dataRaw).then((data) => data);
-    if (result.status.httpCode === 200) {
-      return result;
-    }
-  }
-
-  return false;
-}
+import {
+  fetchLInkListFromCKGApi,
+} from './api-v3-controller.js';
 
 export function findMatchExistingSEOPage(path) {
   const pathMatch = (e) => e.path === path;
